Brandonpatch file
Well, when using ADL.DRV instead of SNDBLAST.DRV, the arcade game chiptune music doesn't sound like chiptunes. This is because the instruments for the arcade game were previously using OPL3-only features which SNDBLAST.DRV takes advantage of but ADL.DRV is OPL2 only and couldn't. So it defaulted to standard features and the result was the instruments sounding nothing like chiptunes. This new patch bank features the chiptune instruments that use OPL2 features to make it compatible with both.